refactor: logger middleware to use centralized pino instance config
