+++ src/config/logger.ts
@@ -1,0 +1,19 @@
+import { pino } from "pino";
+import { isDevelopment } from "@/config/env";
+
+const logger = pino({
+  level: isDevelopment ? "debug" : "info",
+  transport: isDevelopment
+    ? {
+        target: "pino-pretty",
+        options: {
+          colorize: true,
+          translateTime: "SYS:standard",
+          ignore: "pid,hostname",
+        },
+      }
+    : undefined,
+  timestamp: true,
+});
+
+export { logger };
